Transfer type definitions to API layer
